+++ src/main/java/com/board_of_ads/configs/security/UserDetailsServiceImpl.java
@@ -1,7 +1,7 @@
-package com.avito.configs.security;
+package com.board_of_ads.configs.security;
 
-import com.avito.models.User;
-import com.avito.service.interfaces.UserService;
+import com.board_of_ads.models.User;
+import com.board_of_ads.service.interfaces.UserService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.security.core.userdetails.UserDetails;
