+++ src/main/java/com/board_of_ads/configs/security/handler/LoginSuccessHandler.java
@@ -1,4 +1,4 @@
-package com.avito.configs.security.handler;
+package com.board_of_ads.configs.security.handler;
 
 import org.springframework.security.core.Authentication;
 import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
