+++ src/main/java/com/board_of_ads/controllers/rest/kladr/RegionRestController.java
@@ -1,8 +1,8 @@
-package com.avito.controllers.rest.kladr;
+package com.board_of_ads.controllers.rest.kladr;
 
 
-import com.avito.models.kladr.Region;
-import com.avito.service.interfaces.RegionService;
+import com.board_of_ads.models.kladr.Region;
+import com.board_of_ads.service.interfaces.RegionService;
 import lombok.AllArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
