+++ src/main/java/com/board_of_ads/controllers/simple/MainPageController.java
@@ -1,13 +1,11 @@
-package com.avito.controllers.simple;
+package com.board_of_ads.controllers.simple;
 
-import com.avito.models.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
 
 @Controller
 public class MainPageController {
