+++ src/main/java/com/board_of_ads/models/dto/ProfilePostingDTO.java
@@ -1,4 +1,4 @@
-package com.avito.models.dto;
+package com.board_of_ads.models.dto;
 
 import lombok.Data;
