+++ src/main/java/com/board_of_ads/repository/CategoryRepository.java
@@ -1,6 +1,6 @@
-package com.avito.repository;
+package com.board_of_ads.repository;
 
-import com.avito.models.Category;
+import com.board_of_ads.models.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
