+++ src/main/java/com/board_of_ads/repository/PasswordResetTokenRepository.java
@@ -1,7 +1,7 @@
-package com.avito.repository;
+package com.board_of_ads.repository;
 
-import com.avito.models.PasswordResetToken;
-import com.avito.models.User;
+import com.board_of_ads.models.PasswordResetToken;
+import com.board_of_ads.models.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
