+++ src/main/java/com/board_of_ads/repository/PostingRepository.java
@@ -1,7 +1,7 @@
-package com.avito.repository;
+package com.board_of_ads.repository;
 
-import com.avito.models.User;
-import com.avito.models.posting.Posting;
+import com.board_of_ads.models.User;
+import com.board_of_ads.models.posting.Posting;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
