+++ src/main/java/com/board_of_ads/service/interfaces/CityService.java
@@ -1,6 +1,6 @@
-package com.avito.service.interfaces;
+package com.board_of_ads.service.interfaces;
 
-import com.avito.models.kladr.City;
+import com.board_of_ads.models.kladr.City;
 
 
 import java.util.List;
