+++ src/main/java/com/board_of_ads/service/interfaces/SocialNetworkService.java
@@ -1,6 +1,6 @@
-package com.avito.service.interfaces;
+package com.board_of_ads.service.interfaces;
 
-import com.avito.models.User;
+import com.board_of_ads.models.User;
 
 import java.util.Map;
