+++ src/main/java/com/board_of_ads/service/interfaces/UserService.java
@@ -1,9 +1,8 @@
-package com.avito.service.interfaces;
+package com.board_of_ads.service.interfaces;
 
 
-import com.avito.models.PasswordResetToken;
-import com.avito.models.User;
-import com.avito.models.posting.Posting;
+import com.board_of_ads.models.PasswordResetToken;
+import com.board_of_ads.models.User;
 
 import java.util.List;
