+++ src/test/java/com/board_of_ads/AvitoApplicationTests.java
@@ -1,4 +1,4 @@
-package com.avito;
+package com.board_of_ads;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
